pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pro-g-rammingchallenges4"
version = "0.1.0"
description = "Packaging metadata and optional extras for the Pro-g-ramming challenge repository."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "saintwithataint" }]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["challenges", "algorithms", "education", "games"]
dependencies = []

[project.optional-dependencies]
practical = [
    "numpy>=1.26,<2.0",
    "pandas>=2.2,<2.3",
    "requests>=2.31,<3.0",
    "Flask>=3.0,<3.1",
    "pypdf>=4.3,<4.4",
    "mutagen>=1.47,<1.48",
    "cryptography>=42.0,<44.0",
    "tqdm>=4.66,<4.67",
    "boto3>=1.34,<1.35",
    "scipy>=1.11,<1.13",
    "sympy>=1.12,<1.15",
    "opencv-python>=4.10,<5.0",
    "Pillow>=10.4,<10.5",
    "matplotlib>=3.8,<3.9",
    "plotly>=5.22,<5.23",
    "imageio>=2.34,<2.35",
    "pygame>=2.6,<2.7",
    "librosa>=0.10,<0.11",
    "soundfile>=0.12,<0.13",
    "sounddevice>=0.4,<0.5",
    "mido>=1.3,<1.4",
    "python-rtmidi>=1.5,<1.6",
    "PyYAML>=6.0.2,<6.1",
    "pyproj>=3.6.1,<3.7",
    "python-xlib>=0.33,<0.34",
    "colorama>=0.4.6,<0.5",
    "pynput>=1.7,<1.8",
    "pyautogui>=0.9,<0.10",
    "markdown>=3.6,<3.7",
    "tkhtmlview>=0.3.1,<0.4",
]
algorithmic = [
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.13",
    "matplotlib>=3.8,<3.9",
    "Pillow>=10.4,<10.5",
    "requests>=2.31,<3.0",
    "beautifulsoup4>=4.12,<5.0",
    "networkx>=3.3,<3.4",
    "yt-dlp>=2024.7.1",
]
visual = [
    "numpy>=1.26,<2.0",
    "Pillow>=10.4,<10.5",
    "matplotlib>=3.8,<3.9",
    "imageio>=2.34,<2.35",
    "plotly>=5.22,<5.23",
    "colour-science>=0.4,<0.5",
    "vpython>=7.6,<7.7",
    "opencv-python>=4.10,<5.0",
    "sounddevice>=0.4,<0.5",
]
audio = [
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.13",
    "librosa>=0.10,<0.11",
    "soundfile>=0.12,<0.13",
    "sounddevice>=0.4,<0.5",
    "mutagen>=1.47,<1.48",
]
games = [
    "numpy>=1.26,<2.0",
    "matplotlib>=3.8,<3.9",
    "pygame>=2.6,<2.7",

    "PySide6>=6.7,<6.8",

    "tcod>=16.2,<17.0",

]
emulation = [
    "pygame>=2.6,<2.7",
]
constructive_solid_geometry = [
    "scikit-image>=0.24,<0.25",
    "trimesh>=4.4,<4.5",
]
ai = [
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.13",
    "scikit-learn>=1.5,<1.6",
    "torch>=2.5,<2.6",
    "torchvision>=0.18,<0.19",
]
ai_roguelike = [
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.13",
    "tcod>=16.2,<17.0",
]
midi = [
    "mido>=1.3,<1.4",
    "python-rtmidi>=1.5,<1.6",
]
web = [
    "requests>=2.31,<3.0",
    "Flask>=3.0,<3.1",
    "plotly>=5.22,<5.23",
]
security = [
    "cryptography>=42.0,<44.0",
]
documents = [
    "pypdf>=4.3,<4.4",
]
markdown = [
    "markdown>=3.6,<3.7",
    "tkhtmlview>=0.3.1,<0.4",
    "networkx>=3.3,<3.4",
]
desktop = [
    "python-xlib>=0.33,<0.34",
    "colorama>=0.4.6,<0.5",
    "pynput>=1.7,<1.8",
    "pyautogui>=0.9,<0.10",
]
geo = [
    "pyproj>=3.6.1,<3.7",
    "PyYAML>=6.0.2,<6.1",
]
media = [
    "imageio>=2.34,<2.35",
    "yt-dlp>=2024.7.1",
]
developer = [
    "pytest>=8.3,<8.4",
    "ruff>=0.5,<0.6",
    "mypy>=1.11,<1.12",
    "mdformat>=0.7.17,<0.8",
]
all = [
    "numpy>=1.26,<2.0",
    "pandas>=2.2,<2.3",
    "requests>=2.31,<3.0",
    "Flask>=3.0,<3.1",
    "pypdf>=4.3,<4.4",
    "mutagen>=1.47,<1.48",
    "cryptography>=42.0,<44.0",
    "tqdm>=4.66,<4.67",
    "boto3>=1.34,<1.35",
    "scipy>=1.11,<1.13",
    "opencv-python>=4.10,<5.0",
    "Pillow>=10.4,<10.5",
    "plotly>=5.22,<5.23",
    "imageio>=2.34,<2.35",
    "pygame>=2.6,<2.7",
    "librosa>=0.10,<0.11",
    "soundfile>=0.12,<0.13",
    "sounddevice>=0.4,<0.5",
    "mido>=1.3,<1.4",
    "python-rtmidi>=1.5,<1.6",
    "PyYAML>=6.0.2,<6.1",
    "pyproj>=3.6.1,<3.7",
    "python-xlib>=0.33,<0.34",
    "colorama>=0.4.6,<0.5",
    "pynput>=1.7,<1.8",
    "pyautogui>=0.9,<0.10",
    "matplotlib>=3.8,<3.9",
    "beautifulsoup4>=4.12,<5.0",
    "yt-dlp>=2024.7.1",
    "colour-science>=0.4,<0.5",
    "vpython>=7.6,<7.7",
    "scikit-learn>=1.5,<1.6",
    "gymnasium>=0.29,<0.30",
    "stable-baselines3>=2.3,<2.4",
    "torch>=2.5,<2.6",
    "sympy>=1.12,<1.15",
    "markdown>=3.6,<3.7",
    "tkhtmlview>=0.3.1,<0.4",
    "pytest>=8.3,<8.4",
    "ruff>=0.5,<0.6",
    "mypy>=1.11,<1.12",
    "scikit-image>=0.24,<0.25",
    "trimesh>=4.4,<4.5",
]

[project.urls]
Homepage = "https://github.com/saintwithataint/Pro-g-rammingChallenges4"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"pro_g_rammingchallenges4.emulation.ti86" = ["opcode_truth.json"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
